corrections in the entitys and the  file persistence.xml

+++ src/br/com/ServerRouteSOAP/model/AbstractEntity.java
@@ -1,7 +1,6 @@
 package br.com.ServerRouteSOAP.model;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 
 /**
  * Estipula um contrato base para as entidades persistentes da aplicação.
@@ -15,6 +14,6 @@ public interface AbstractEntity extends Serializable {
 	 * @return A referência para a chave primária (Primary Key) de cada objeto persistido.
 	 * 		   Caso o objeto ainda não tenha sido persistido, deve retornar <code>null</code>.
 	 */
-	public BigInteger getId();
+	public Long getId();
 	
 }
