+++ src/br/com/ServerRouteSOAP/model/Aresta.java
@@ -1,7 +1,6 @@
 package br.com.ServerRouteSOAP.model;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
@@ -13,13 +12,13 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 
 @Entity
-public class Aresta implements AbstractEntity, Serializable{
+public class Aresta implements AbstractEntity , Serializable{
 
 	private static final long serialVersionUID = 1L;
-
+	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private BigInteger id;
+	private Long id;
 	@Column
 	private int peso;
 	
@@ -41,11 +40,11 @@ public class Aresta implements AbstractEntity, Serializable{
             this.destino = v2;
     }
     
-    public BigInteger getId() {
+    public Long getId() {
 		return id;
 	}
 
-	public void setId(BigInteger id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
